+++ BoxOffice/Network/NetworkRequest.swift
@@ -1,0 +1,16 @@
+//
+//  NetworkRequest.swift
+//  BoxOffice
+//
+//  Created by Effie on 2/19/24.
+//
+
+struct NetworkRequest {
+  
+}
+
+//extension NetworkRequest: URLRequestConvertible {
+//  func toURLRequest() -> URLRequest {
+//    return .init(url: .init(string: "") )
+//  }
+//}
