+++ BoxOffice/Network/URLRequestConvertible.swift
@@ -1,0 +1,12 @@
+//
+//  URLRequestConvertible.swift
+//  BoxOffice
+//
+//  Created by Effie on 2/19/24.
+//
+
+import Foundation
+
+protocol URLRequestConvertible {
+  func toURLRequest() -> URLRequest
+}
